+++ packages/SequenceTranslator/src/package.ts
@@ -17,16 +17,16 @@ import {getExternalAppViewFactories} from './plugins/mermade';
 import {defaultErrorHandler} from './utils/err-info';
 
 //polytool specific
-import {getPolyToolConversionDialog, polyToolEnumerateChemUI} from './polytool/pt-dialog';
+import {polyToolConvert, polyToolConvertUI} from './polytool/pt-dialog';
+import {polyToolEnumerateChemUI} from './polytool/pt-dialog';
 import {polyToolEnumerateHelmUI} from './polytool/pt-enumeration-helm-dialog';
 import {_setPeptideColumn} from './polytool/utils';
 import {PolyToolCsvLibHandler} from './polytool/csv-to-json-monomer-lib-converter';
 import {ITranslationHelper} from './types';
 import {addContextMenuUI} from './utils/context-menu';
 import {PolyToolConvertFuncEditor} from './polytool/pt-convert-editor';
-import {doPolyToolConvert} from './polytool/pt-conversion';
 
-export const _package: OligoToolkitPackage = new OligoToolkitPackage(/*{debug: true}/**/);
+export const _package: OligoToolkitPackage = new OligoToolkitPackage({debug: true}/**/);
 
 //name: Oligo Toolkit
 //meta.icon: img/icons/toolkit.png
@@ -158,14 +158,8 @@ async function getSpecifiedAppView(appName: string): Promise<DG.ViewBase> {
 //top-menu: Bio | PolyTool | Convert...
 //name: polyToolConvert
 //description: Perform cyclization of polymers
-export async function polyToolConvert(): Promise<void> {
-  let dialog: DG.Dialog;
-  try {
-    dialog = await getPolyToolConversionDialog();
-    dialog.show();
-  } catch (err: any) {
-    grok.shell.warning('To run PolyTool Conversion, open a dataframe with macromolecules');
-  }
+export async function polyToolConvertTopMenu(): Promise<void> {
+  await polyToolConvertUI();
 }
 
 //name: getPolyToolConvertEditor
@@ -188,7 +182,7 @@ export async function getPolyToolConvertEditor(call: DG.FuncCall): Promise<DG.Co
 export async function polyToolConvert2(table: DG.DataFrame,
   seqCol: DG.Column, generateHelm: boolean, chiralityEngine: boolean, rules: string[]
 ): Promise<DG.Column<string>> {
-  const ptConvertRes = await doPolyToolConvert(seqCol, generateHelm, rules, chiralityEngine);
+  const ptConvertRes = await polyToolConvert(seqCol, generateHelm, chiralityEngine, rules);
   return ptConvertRes[0];
 }
 
@@ -196,14 +190,14 @@ export async function polyToolConvert2(table: DG.DataFrame,
 //top-menu: Bio | PolyTool | Enumerate HELM...
 //name: polyToolEnumerateHelm
 //description: Perform cyclization of polymers
-export async function polyToolEnumerateHelm(): Promise<void> {
+export async function polyToolEnumerateHelmTopMenu(): Promise<void> {
   await polyToolEnumerateHelmUI(grok.shell.tv?.dataFrame.currentCell);
 }
 
 //top-menu: Bio | PolyTool | Enumerate Chem...
 //name: polyToolEnumerateChem
 //description: Perform cyclization of polymers
-export async function polyToolEnumerateChem(): Promise<void> {
+export async function polyToolEnumerateChemTopMenu(): Promise<void> {
   polyToolEnumerateChemUI();
 }