+++ js-api/src/widgets.ts
@@ -1236,7 +1236,7 @@ export class InputForm extends DartWrapper {
 
   getInput(propertyName: string): InputBase { return toJs(api.grok_InputForm_GetInput(this.dart, propertyName)); }
 
-  get source(): any { return api.grok_InputForm_Get_Source(this.dart); };
+  get source(): any { return toJs(api.grok_InputForm_Get_Source(this.dart)); };
 
   set source(source: any) { api.grok_InputForm_Set_Source(this.dart, toDart(source)); };